Created Modal for Messages table and made a few minor fixes

+++ client/src/pages/Messages.jsx
@@ -1,9 +1,14 @@
 import { useDispatch, useSelector } from "react-redux";
 import { fetchMessageAsyncThunk, deleteMessageAsyncThunk } from "../redux/slice/messageSlice";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
+
+import { Modal } from "../components/Modal";
 
 export const Messages = () => {
     const { messageList } = useSelector((state) => state.message);
+    const [showModal, setShowModal] = useState(false);
+    const [messageId, setMessageId] = useState(null);
+
     const dispatch = useDispatch();
 
     useEffect(() => {
@@ -11,69 +16,71 @@ export const Messages = () => {
     }, []);
 
     const handleDelete = (id) => {
+        const confirm = window.confirm("Are you sure you want to delete this message?");
+        if (!confirm) return;
         dispatch(deleteMessageAsyncThunk(id));
-    }
+    };
+
+    const handleView = (id) => {
+        setMessageId(id);
+        setShowModal(true);
+    };
 
     return (
         <main className="my-5">
             <div className="container px-4 px-lg-5">
                 <div className="row gx-4 gx-lg-5 justify-content-center">
                     <h1 className="text-center mb-5">Your Messages</h1>
-                    <div className="scrollable-table">
-                        {
-                            messageList.length > 0 ? (
+                    {
+                        messageList.length > 0 ? (
+                            <div className="scrollable-table">
                                 <table className="table">
                                     <thead>
                                         <tr>
-                                            <th className="px-0 py-4" scope="col">İsim</th>
-                                            <th className="px-0 py-4" scope="col">Email</th>
-                                            <th className="px-0 py-4" scope="col">Gönderilme Tarihi</th>
-                                            <th className="px-0 py-4" scope="col">Mesaj</th>
-                                            <th className="px-0 py-4" scope="col">İşlem</th>
+                                            <th className="px-0 py-2" scope="col">İsim</th>
+                                            <th className="px-0 py-2" scope="col">Email</th>
+                                            <th className="px-0 py-2" scope="col">Gönderilme Tarihi</th>
+                                            <th className="px-0 py-2" scope="col">Mesaj</th>
+                                            <th className="px-0 py-2 text-end" scope="col">İşlem</th>
                                         </tr>
                                     </thead>
                                     <tbody>
                                         {
                                             messageList.map((message) => (
-                                                <tr key={message._id} className="align-middle" >
+                                                <tr key={message._id} className="body-row align-middle" >
                                                     <td>{message.name}</td>
                                                     <td>{message.email}</td>
                                                     <td>{message.createdAt}</td>
                                                     <td>
-                                                        <button type="button" className="btn btn-success btn-sm" data-bs-toggle="modal" data-bs-target={`#message_modal_${message._id}`}>
+                                                        <button type="button" className="btn btn-success btn-sm" onClick={() => handleView(message._id)}>
                                                             <i className="bi bi-eye"></i>
                                                         </button>
-                                                        <div className="modal fade" id={`message_modal_${message._id}`} tabIndex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
-                                                            <div className="modal-dialog modal-dialog-centered">
-                                                                <div className="modal-content">
-                                                                    <div className="modal-header">
-                                                                        <h1 className="modal-title fs-5" id="exampleModalLabel">Message</h1>
-                                                                        <button type="button" className="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-                                                                    </div>
-                                                                    <div className="modal-body">
-                                                                        {message.message}
-                                                                    </div>
-                                                                </div>
-                                                            </div>
-                                                        </div>
                                                     </td>
-                                                    <td>
+                                                    <td className="text-end">
                                                         <button className="btn btn-danger btn-sm" onClick={() => handleDelete(message._id)}><i className="bi bi-trash"></i></button>
                                                     </td>
                                                 </tr>
                                             ))}
                                     </tbody>
                                 </table>
-                            ) : (
-                                <div className="alert alert-warning" role="alert">
-                                    <h4 className="alert-heading">No Message</h4>
-                                    <p>There is no message.</p>
-                                </div>
-                            )
-                        }
-                    </div>
+                            </div>
+
+                        ) : (
+                            <div className="alert alert-warning" role="alert">
+                                <h4 className="alert-heading">No Message</h4>
+                                <p>There is no message.</p>
+                            </div>
+                        )
+                    }
                 </div>
             </div>
+            {
+                showModal && (
+                    <Modal show={showModal} onClose={() => setShowModal(false)} modalTitle={"Message"}>
+                        <p>{messageList.find(msg => msg._id === messageId)?.message}</p>
+                    </Modal>
+                )
+            }
         </main>
     );
 }
